+++ src/main/java/com/tdsi/sn/app_moblile_api/Services/Verification.java
@@ -30,8 +30,8 @@ public class Verification {
             e.printStackTrace();
         }
     }
-    public BigInteger getPublic(){
-        return pub.getPublicExponent();
+    public Long  getPublic(){
+        return pub.getPublicExponent().longValue();
     }
     public BigInteger decodd(BigInteger password){
           return password.modPow(priv.getPrivateExponent(),priv.getModulus());
